Improved: replaced order add/update item API with transfer order APIs (#1302)
